/api/constants to /constants
