+++ App.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import { StatusBar } from 'react-native';
 import { AppLoading } from 'expo';
-import { cacheImages } from './src/api/util';
+import { func } from './src/constants';
 
 import Stack from './src/navigation/Stack';
-
-import preloadImages from './src/api/preloadImages';
 
 export default class App extends React.Component {
   constructor(props) {
@@ -14,16 +12,6 @@ export default class App extends React.Component {
     this.state = {
       isLoading: true
     };
-
-    this.loadAssetsAsync = this.loadAssetsAsync.bind(this);
-  }
-
-  async loadAssetsAsync() {
-    const imageAssets = cacheImages(preloadImages);
-
-    await Promise.all([...imageAssets]).then(() => {
-      this.setState({ isLoading: false });
-    });
   }
 
   render() {
@@ -33,7 +21,7 @@ export default class App extends React.Component {
       return (
         <AppLoading
           onFinish={() => this.setState({ isLoading: false })}
-          startAsync={this.loadAssetsAsync}
+          startAsync={func.loadAssetsAsync}
         />
       );
     }
