--- src/api/functions.js
@@ -1,5 +1,0 @@
-import registerPushNotifcation from './functions/registerForPushNotificationsAsync';
-
-export default {
-  registerPushNotifcation
-};
