+++ src/components/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Image, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import PropTypes from 'prop-types';
-import { colors, device, fonts, images } from '../api/constants';
+import { colors, device, fonts, images } from '../constants';
 
 import SvgArrowLeft from './icons/Svg.ArrowLeft';
