+++ src/constants/device.js
