+++ src/constants/functions.js
@@ -1,0 +1,71 @@
+import { Image } from 'react-native';
+import { Asset, Font, Permissions } from 'expo';
+
+import preloadFonts from './preloadFonts';
+import preloadImages from './preloadImages';
+
+// cache fonts
+// /////////////////////////////////////////////////////////////////////////////
+const cacheFonts = fonts => fonts.map(font => Font.loadAsync(font));
+
+// cache images
+// /////////////////////////////////////////////////////////////////////////////
+const cacheImages = images => {
+  const imagesArray = Object.values(images);
+
+  return imagesArray.map(image => {
+    if (typeof image === 'string') {
+      return Image.prefetch(image);
+    }
+
+    return Asset.fromModule(image).downloadAsync();
+  });
+};
+
+// preload async
+// /////////////////////////////////////////////////////////////////////////////
+const loadAssetsAsync = async () => {
+  // preload assets
+  const fontAssets = cacheFonts(preloadFonts);
+  const imageAssets = cacheImages(preloadImages);
+
+  // promise load all
+  return Promise.all([...fontAssets, ...imageAssets]);
+};
+
+// camera permissions
+// /////////////////////////////////////////////////////////////////////////////
+const cameraAccessAsync = async () => {
+  // get exisiting camera permissions first
+  const { status: existingStatus } = await Permissions.getAsync(
+    Permissions.CAMERA
+  );
+  let finalStatus = existingStatus;
+
+  // ask again to grant camera permissions (if not already allowed)
+  if (existingStatus !== 'granted') {
+    const { status } = await Permissions.askAsync(Permissions.CAMERA);
+    finalStatus = status;
+  }
+
+  return finalStatus === 'granted';
+};
+
+// format seconds
+// /////////////////////////////////////////////////////////////////////////////
+const formatTime = sec => {
+  const padTime = (num, size) => `000${num}`.slice(size * -1);
+  const time = parseFloat(sec).toFixed(3);
+  const minutes = Math.floor(time / 60) % 60;
+  const seconds = Math.floor(time - minutes * 60);
+
+  return `${padTime(minutes, 1)}:${padTime(seconds, 2)}`;
+};
+
+export default {
+  cacheFonts,
+  cacheImages,
+  loadAssetsAsync,
+  cameraAccessAsync,
+  formatTime
+};
