+++ src/constants/globalStyles.js
