+++ src/constants/index.js
@@ -1,13 +1,8 @@
 import colors from './colors';
 import device from './device';
 import fonts from './fonts';
-import functions from './functions';
+import func from './functions';
 import gStyle from './globalStyles';
 import images from './preloadImages';
 
-export { colors };
-export { device };
-export { fonts };
-export { functions };
-export { gStyle };
-export { images };
+export { colors, device, fonts, func, gStyle, images };
