+++ src/navigation/defaultOptions.js
@@ -1,4 +1,4 @@
-import { gStyle } from '../api/constants';
+import { gStyle } from '../constants';
 
 export default () => ({
   headerStyle: gStyle.navHeaderStyle
