+++ src/screens/Downloads.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import PropTypes from 'prop-types';
-import { colors, fonts, gStyle } from '../api/constants';
+import { colors, fonts, gStyle } from '../constants';
 
 import Cast from '../components/Cast';
 import Header from '../components/Header';
