+++ src/screens/ModalAddProfile.js
@@ -9,7 +9,7 @@ import {
   View
 } from 'react-native';
 import PropTypes from 'prop-types';
-import { colors, fonts, gStyle, images } from '../api/constants';
+import { colors, fonts, gStyle, images } from '../constants';
 
 import HeaderManage from '../components/HeaderManage';
