+++ src/screens/ModalVideo.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 import { ScreenOrientation } from 'expo';
-import { gStyle } from '../api/constants';
+import { gStyle } from '../constants';
 
 class ModalVideo extends React.Component {
   componentDidMount() {
