+++ src/screens/Movies.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 import PropTypes from 'prop-types';
-import { gStyle } from '../api/constants';
+import { gStyle } from '../constants';
 
 import Cast from '../components/Cast';
 import HeaderHome from '../components/HeaderHome';
